+++ lib/asherah.rb
@@ -44,6 +44,8 @@ module Asherah
     # @yield [Config]
     # @return [void]
     def configure
+      raise Asherah::Error::AlreadyInitialized if @initialized
+
       config = Config.new
       yield config
       config.validate!
@@ -53,6 +55,7 @@ module Asherah
 
       result = SetupJson(config_buffer)
       Error.check_result!(result, 'SetupJson failed')
+      @initialized = true
     end
 
     # Encrypts data for a given partition_id and returns DataRowRecord in JSON format.
@@ -104,7 +107,10 @@ module Asherah
 
     # Stop the Asherah instance
     def shutdown
+      raise Asherah::Error::NotInitialized unless @initialized
+
       Shutdown()
+      @initialized = false
     end
 
     private